Add function to compute adj matrix for all mf->grc and gl->grc conns.
The function is called AmatrixMfsGloms, and unlike Amatrix it doesn't
try to ignore unconnected mossy fibers. Plus it also calculates the
adjacency matrix for the network composed by glomeruli and granules,
so that this matrix can be used together with the information on the
glomerular locations to build a spatial graph.